update module Categories. Fix the test grandle.

+++ business/src/main/java/com/github/mytravelsapp/business/interactor/RemoveCategoryInteractor.java
@@ -1,5 +1,6 @@
 package com.github.mytravelsapp.business.interactor;
 
+import com.github.mytravelsapp.business.exception.ExistCategoryInTravelPlacesBusinessException;
 import com.github.mytravelsapp.business.executor.PostExecutionThread;
 import com.github.mytravelsapp.business.executor.ThreadExecutor;
 import com.github.mytravelsapp.business.repository.CategoryRepository;
@@ -14,16 +15,22 @@ public class RemoveCategoryInteractor extends AbstractBackgroundInteractor<Boole
 
     private final CategoryRepository categoryRepository;
 
+    private final TravelPlacesRepository travelPlacesRepository;
+
     private long categoryId;
 
     @Inject
-    public RemoveCategoryInteractor(final ThreadExecutor pThreadExecutor, final PostExecutionThread pPostExecutionThread, final CategoryRepository pCategoryRepository) {
+    public RemoveCategoryInteractor(final ThreadExecutor pThreadExecutor, final PostExecutionThread pPostExecutionThread, final CategoryRepository pCategoryRepository, final TravelPlacesRepository pTravelPlacesRepository) {
         super(pThreadExecutor, pPostExecutionThread);
         this.categoryRepository = pCategoryRepository;
+        this.travelPlacesRepository = pTravelPlacesRepository;
     }
 
     @Override
     public Boolean backgroundTask() throws Exception {
+       if(!travelPlacesRepository.findByIdCategory(getCategoryId()).isEmpty()){
+           throw new ExistCategoryInTravelPlacesBusinessException("Category with travel places associated!");
+       }
         categoryRepository.remove(getCategoryId());
         return Boolean.TRUE;
     }
